requirements.txt
packaging @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_a6_qk3qyg7/croot/packaging_1734472142254/work
pillow @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_85xj2lenf1/croot/pillow_1738010251686/work
pytesseract @ file:///Users/builder/cbouss/buildout/croot/pytesseract_1739490350820/work
setuptools==75.8.0
wheel==0.45.1
